+++ core/di/src/main/java/com/lanpet/core/di/ApiModule.kt
@@ -3,8 +3,8 @@ package com.lanpet.core.di
 import com.lanpet.core.manager.AuthStateHolder
 import com.lanpet.data.service.AccountApiClient
 import com.lanpet.data.service.AccountApiService
-import com.lanpet.data.service.AuthApiClient
-import com.lanpet.data.service.BaseApiService
+import com.lanpet.data.service.ProfileApiClient
+import com.lanpet.data.service.ProfileApiService
 import com.lanpet.data.service.FreeBoardApiClient
 import com.lanpet.data.service.FreeBoardApiService
 import dagger.Module
@@ -21,18 +21,18 @@ object ApiModule {
     @Provides
     fun provideBaseApiService(
         @Named("BaseApiUrl") baseApiUrl: String,
-    ): BaseApiService = AuthApiClient(baseApiUrl).getService()
+    ): ProfileApiService = ProfileApiClient(baseApiUrl).getService()
 
     @Singleton
     @Provides
     fun provideAuthApiClient(
         @Named("BaseApiUrl") baseApiUrl: String,
         authStateHolder: AuthStateHolder,
-    ): AuthApiClient = AuthApiClient(baseApiUrl)
+    ): ProfileApiClient = ProfileApiClient(baseApiUrl)
 
     @Singleton
     @Provides
-    fun provideAuthApiService(authApiClient: AuthApiClient): BaseApiService = authApiClient.getService()
+    fun provideAuthApiService(authApiClient: ProfileApiClient): ProfileApiService = authApiClient.getService()
 
     @Singleton
     @Provides
