--- data/service/src/main/java/com/lanpet/data/service/BaseApiService.kt
@@ -1,15 +1,0 @@
-package com.lanpet.data.service
-
-import retrofit2.http.GET
-import retrofit2.http.POST
-
-interface BaseApiService {
-    @POST("profiles")
-    suspend fun createProfile()
-
-    @GET("profiles")
-    suspend fun getProfileList()
-
-    @GET("profiles/{id}")
-    suspend fun getProfileDetail(id: String)
-}
